fix for self test error reported about mycurl_easyhandle_node.cpp line 176

+++ src/libdar/i_entrepot_libcurl.cpp
@@ -351,7 +351,7 @@ namespace libdar
 	if(auth_from_file)
 	{
 	    easyh.setopt_global(CURLOPT_USERNAME, real_login);
-	    easyh.setopt_global(CURLOPT_NETRC, CURL_NETRC_OPTIONAL);
+	    easyh.setopt_global(CURLOPT_NETRC, (long)(CURL_NETRC_OPTIONAL));
 	}
 	else // login + password authentication
 	{
